FLyttet tasks og service tilknyttet endring og oppgave til egen mappe
