+++ src/main/kotlin/no/nav/familie/ef/personhendelse/forvaltning/BehandleAutomatiskInntektsendringForvaltningsController.kt
@@ -2,9 +2,9 @@ package no.nav.familie.ef.personhendelse.forvaltning
 
 import io.swagger.v3.oas.annotations.Operation
 import no.nav.familie.ef.personhendelse.client.SakClient
-import no.nav.familie.ef.personhendelse.inntekt.InntektOppgaveService
-import no.nav.familie.ef.personhendelse.inntekt.InntektsendringerService
 import no.nav.familie.ef.personhendelse.inntekt.LoggInntektForPersonTask
+import no.nav.familie.ef.personhendelse.inntekt.endring.InntektsendringerService
+import no.nav.familie.ef.personhendelse.inntekt.oppgave.InntektOppgaveService
 import no.nav.familie.prosessering.internal.TaskService
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import org.springframework.web.bind.annotation.GetMapping
