+++ src/main/kotlin/no/nav/familie/ef/personhendelse/inntekt/InntektsendringerScheduler.kt
@@ -1,5 +1,7 @@
 package no.nav.familie.ef.personhendelse.inntekt
 
+import no.nav.familie.ef.personhendelse.inntekt.endring.InntektsendringerService
+import no.nav.familie.ef.personhendelse.inntekt.oppgave.InntektOppgaveService
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.scheduling.annotation.Scheduled
