+++ src/main/kotlin/no/nav/familie/ef/personhendelse/inntekt/oppgave/FinnPersonerFyltTjueFemArbeidsavklaringspengerTask.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.ef.personhendelse.inntekt
+package no.nav.familie.ef.personhendelse.inntekt.oppgave
 
 import com.fasterxml.jackson.module.kotlin.readValue
 import no.nav.familie.ef.personhendelse.client.pdl.PdlClient
