+++ src/main/kotlin/no/nav/familie/ef/personhendelse/inntekt/oppgave/OpprettOppgaverForUføretrygdsendringerTask.kt
@@ -1,6 +1,5 @@
-package no.nav.familie.ef.personhendelse.inntekt
+package no.nav.familie.ef.personhendelse.inntekt.oppgave
 
-import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
 import no.nav.familie.kontrakter.felles.objectMapper
 import no.nav.familie.prosessering.AsyncTaskStep
@@ -11,42 +10,43 @@ import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import java.time.YearMonth
 import java.util.Properties
-import kotlin.collections.set
 
 @Service
 @TaskStepBeskrivelse(
-    taskStepType = OpprettOppgaverForNyeVedtakUføretrygdTask.TYPE,
+    taskStepType = OpprettOppgaverForUføretrygdsendringerTask.TYPE,
     maxAntallFeil = 1,
     settTilManuellOppfølgning = true,
-    beskrivelse = "Oppretter oppgave for nye uføretrygd-vedtak på person",
+    beskrivelse = "Oppretter oppgave for uføretrygsendringer på person",
 )
-class OpprettOppgaverForNyeVedtakUføretrygdTask(
+class OpprettOppgaverForUføretrygdsendringerTask(
     private val inntektOppgaveService: InntektOppgaveService,
 ) : AsyncTaskStep {
+    val logger: Logger = LoggerFactory.getLogger(this::class.java)
     val secureLogger: Logger = LoggerFactory.getLogger("secureLogger")
 
     override fun doTask(task: Task) {
-        val personIdent = objectMapper.readValue<PayloadOpprettOppgaverForNyeVedtakUføretrygdTask>(task.payload).personIdent
-        secureLogger.info("Oppretter oppgaver for nye vedtak uføretrygd for person: $personIdent")
-        inntektOppgaveService.opprettOppgaveForInntektsendring(personIdent, inntektOppgaveService.lagOppgavetekstVedNyYtelseUføretrygd())
+        val personIdent = objectMapper.readValue<PayloadOpprettOppgaverForUføretrygdsendringerTask>(task.payload).personIdent
+        secureLogger.info("Oppretter oppgaver for uføretrygdsendringer ${task.payload}")
+        inntektOppgaveService.opprettOppgaveForUføretrygdEndring(personIdent, inntektOppgaveService.lagOppgavetekstVedEndringUføretrygd(YearMonth.now().minusMonths(1)))
     }
 
     companion object {
-        const val TYPE = "opprettOppgaverForNyeVedtakUføretrygdTask"
+        const val TYPE = "opprettOppgaverForUføretrygdsendringerTask"
 
-        fun opprettTask(payload: PayloadOpprettOppgaverForNyeVedtakUføretrygdTask): Task =
+        fun opprettTask(payload: PayloadOpprettOppgaverForUføretrygdsendringerTask): Task =
             Task(
                 type = TYPE,
                 payload = objectMapper.writeValueAsString(payload),
                 properties =
                     Properties().apply {
                         this["personIdent"] = payload.personIdent
+                        this["årMåned"] = payload.årMåned.toString()
                     },
             )
     }
 }
 
-data class PayloadOpprettOppgaverForNyeVedtakUføretrygdTask(
+data class PayloadOpprettOppgaverForUføretrygdsendringerTask(
     val personIdent: String,
-    val prosessertYearMonth: YearMonth,
+    val årMåned: YearMonth,
 )
