+++ src/test/kotlin/no/nav/familie/ef/personhendelse/inntekt/InntektOgVedtakEndringTest.kt
@@ -1,5 +1,6 @@
 package no.nav.familie.ef.personhendelse.inntekt
 
+import no.nav.familie.ef.personhendelse.inntekt.endring.BeregningResultat
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import java.time.LocalDateTime
