+++ src/test/kotlin/no/nav/familie/ef/personhendelse/inntekt/OpprettOppgaverForUføretrygdsendringerTaskTest.kt
@@ -3,7 +3,9 @@ package no.nav.familie.ef.personhendelse.inntekt
 import io.mockk.mockk
 import io.mockk.verify
 import no.nav.familie.ef.personhendelse.IntegrasjonSpringRunnerTest
-import no.nav.familie.kontrakter.felles.objectMapper
+import no.nav.familie.ef.personhendelse.inntekt.oppgave.InntektOppgaveService
+import no.nav.familie.ef.personhendelse.inntekt.oppgave.OpprettOppgaverForUføretrygdsendringerTask
+import no.nav.familie.ef.personhendelse.inntekt.oppgave.PayloadOpprettOppgaverForUføretrygdsendringerTask
 import no.nav.familie.prosessering.internal.TaskService
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
